implemented service componenets

+++ src/main/java/org/example/model/Greeting.java
@@ -1,22 +1,25 @@
 package org.example.model;
 
-import java.math.BigInteger;
+
+import javax.persistence.Entity;
 
 /**
  * Created by sergeikostin on 3/1/16.
  */
+
+
 public class Greeting {
-    private BigInteger id;
+    private Long id;
     private String text;
 
     public Greeting() {
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
